feat: restructure array summary
