+++ app/components/formElements/arraySummary/__test__/ArraySummaryDataItems.test.tsx
@@ -23,9 +23,12 @@ const translations = {
   "unterhaltszahlungen.surname": "Nachname",
 };
 
-vi.mock("~/components/arraySummary/ArraySummaryItemButton", () => ({
-  default: () => <div>Mock ArraySummaryItemButton</div>,
-}));
+vi.mock(
+  "~/components/formElements/arraySummary/ArraySummaryItemButton",
+  () => ({
+    default: () => <div>Mock ArraySummaryItemButton</div>,
+  }),
+);
 
 describe("ArraySummaryDataItems", () => {
   afterEach(() => {
